[catbrain] Got more programs working
A pretty fun language

Thinking about constraining it to be data-dependent.

It can be for extracting data, and generating data proportional to 'seq
n'.

You can do a fixed number of columns, but not an arbitrary number?

You can't loop over numbers like seq $n?  The output must be
proportional to the (input, program size).